Simplify basic viewport extensions
