+++ extensions/ohif-dicom-html-extension/src/index.js
@@ -1,3 +1,19 @@
-import OHIFDicomHtmlExtension from './OHIFDicomHtmlExtension.js';
+import OHIFDicomHtmlSopClassHandler from './OHIFDicomHtmlSopClassHandler.js';
+import OHIFDicomHtmlViewport from './OHIFDicomHtmlViewport.js';
 
-export default OHIFDicomHtmlExtension;
+export default class OHIFDicomHtmlExtension {
+  /**
+   * Extension ID is a unique id, might be used for namespacing extension specific redux actions/reducers (?)
+   */
+  getExtensionId() {
+    return 'html';
+  }
+
+  getViewportModule() {
+    return OHIFDicomHtmlViewport;
+  }
+
+  getSopClassHandler() {
+    return OHIFDicomHtmlSopClassHandler;
+  }
+}
