+++ extensions/ohif-dicom-microscopy-extension/src/index.js
@@ -1,3 +1,19 @@
-import OHIFDicomMicroscopyExtension from './OHIFDicomMicroscopyExtension.js';
+import DicomMicroscopySopClassHandler from './DicomMicroscopySopClassHandler.js';
+import DicomMicroscopyViewport from './DicomMicroscopyViewport.js';
 
-export default OHIFDicomMicroscopyExtension;
+export default class OHIFDicomMicroscopyExtension {
+  /**
+   * Extension ID is a unique id, might be used for namespacing extension specific redux actions/reducers (?)
+   */
+  getExtensionId() {
+    return 'microscopy';
+  }
+
+  getViewportModule() {
+    return DicomMicroscopyViewport;
+  }
+
+  getSopClassHandlerModule() {
+    return DicomMicroscopySopClassHandler;
+  }
+}
