+++ extensions/ohif-dicom-pdf-extension/src/index.js
@@ -1,3 +1,19 @@
-import OHIFDicomPDFExtension from './OHIFDicomPDFExtension.js';
+import OHIFDicomPDFSopClassHandler from './OHIFDicomPDFSopClassHandler.js';
+import OHIFDicomPDFViewport from './OHIFDicomPDFViewport.js';
 
-export default OHIFDicomPDFExtension;
+export default class OHIFDicomPDFExtension {
+  /**
+   * Extension ID is a unique id, might be used for namespacing extension specific redux actions/reducers (?)
+   */
+  getExtensionId() {
+    return 'pdf';
+  }
+
+  getViewportModule() {
+    return OHIFDicomPDFViewport;
+  }
+
+  getSopClassHandler() {
+    return OHIFDicomPDFSopClassHandler;
+  }
+}
